Full recode
Don't use a Discord bot at all anymore because it's too much of a hassle.
It's now just a program that sends webhook messages, adding and removing
users is something you have to do yourself, either by editing the tracking list
or coding a discord bot/whatever else to edit it for you.